refactor(commands): split builtin-commands.ts into modular files (Issue #696)
Split the 955-line builtin-commands.ts into smaller, focused modules:

- session-commands.ts: ResetCommand, StatusCommand, HelpCommand (~70 lines)
- node-commands.ts: ListNodesCommand, SwitchNodeCommand, RestartCommand (~85 lines)
- group-commands.ts: Group-related commands + PassiveCommand (~300 lines)
- debug-commands.ts: SetDebugCommand, ShowDebugCommand, ClearDebugCommand (~85 lines)
- schedule-command.ts: ScheduleCommand (~180 lines)
- task-command.ts: TaskCommand (~165 lines)

Benefits:
- Improved code organization and maintainability
- Easier to locate and modify specific commands
- Better adherence to single responsibility principle
- Simplified testing and code review

All 41 command-related tests pass.
